requirements.txt
# Core dependencies from setup.py
click==8.3.0
faker==37.8.0
rdflib==7.2.1
pyyaml==6.0.2
owlrl==7.1.4
jpype1==1.6.0

# Additional dependencies discovered during installation
art==6.5
numpy==2.3.3
owlready2==0.48
tqdm==4.67.1
setuptools==80.9.0
tabulate==0.9.0

# Supporting dependencies (automatically installed with above packages)
colorama==0.4.6
pyparsing==3.2.5
packaging==25.0
tzdata==2025.2

# Install the project itself in editable mode
-e .







